lara/state/jump: port M_Compress to TRX
