+++ src/libtrx/meson.build
@@ -185,6 +185,7 @@ sources = [
   'game/lara/state.c',
   'game/lara/state/climb.c',
   'game/lara/state/extra.c',
+  'game/lara/state/jump.c',
   'game/lara/state/swim.c',
   'game/level/common.c',
   'game/level/faces.c',
